extract the product utility functions into an index file
This reduce the amount of clutter in the utils file and centralize
all the product related functions in one place. This will also make it
easier to add new product related functions in the future without
having to modify the utils file or import multiple files.
