+++ frontend/tests/jest/utils.tsx
@@ -68,24 +68,6 @@ export const renderWithAppContext = (
    options?: RenderOptions
 ): RenderResult => render(ui, { wrapper: GlobalContextProvider, ...options });
 
-/*
- * Utility functions for mocking product data
- */
-export const getMockProduct = (overrides: Partial<Product> = {}): Product => {
-   return {
-      ...mockedProduct,
-      ...overrides,
-   };
-};
-
-export const getMockProductVariant = (
-   overrides: Partial<ProductVariant> = {}
-): ProductVariant => {
-   return {
-      ...mockedProductVariant,
-      ...overrides,
-   };
-};
 
 export const getMockProductSearchHit = (
    overrides: Partial<ProductSearchHit> = {}
@@ -94,92 +76,6 @@ export const getMockProductSearchHit = (
       ...mockedProductSearchHit,
       ...overrides,
    };
-};
-
-/*
- * To match a real use case, you will need to set the selectedVariant to the
- * first variant in the variants array from the returned product.
- */
-export const getDiscountedProduct = (
-   originalPrice = 49.99,
-   discountPercentage = 10
-) => {
-   const discountedPrice =
-      originalPrice - originalPrice * (discountPercentage / 100);
-
-   const product = getMockProduct({
-      variants: [
-         getMockProductVariant({
-            price: {
-               amount: discountedPrice,
-               currencyCode: CurrencyCode.Usd,
-            },
-            compareAtPrice: {
-               amount: originalPrice,
-               currencyCode: CurrencyCode.Usd,
-            },
-            isDiscounted: true,
-            discountPercentage: discountPercentage,
-         }),
-      ],
-   });
-   return product;
-};
-
-export const getNonDiscountedProduct = (originalPrice = 49.99) => {
-   const product = getMockProduct({
-      variants: [
-         getMockProductVariant({
-            price: {
-               amount: originalPrice,
-               currencyCode: CurrencyCode.Usd,
-            },
-            compareAtPrice: null,
-            isDiscounted: false,
-            discountPercentage: 0,
-         }),
-      ],
-   });
-   return product;
-};
-
-export const getProductOfType = (type: 'battery' | 'tool' | 'part') => {
-   switch (type) {
-      case 'battery':
-         return { ...mockedBatteryProduct };
-      case 'tool':
-         return { ...mockedToolProduct };
-      case 'part':
-         return { ...mockedPartProduct };
-   }
-};
-
-/*
- * These warranty values are listed in iFixit's product warranty policy:
- * https://www.ifixit.com/Info/Warranty#Section_iFixit_Guarantees
- *
- * Full Guarantee refers to Lifetime Warranty.
- * Limited Guarantee refers to 1-Year Warranty.
- * As-Is refers to No Warranty.
- */
-export const getProductWithWarranty = (
-   guarantee: 'full' | 'limited' | 'as-is'
-) => {
-   enum Warranty {
-      'full' = 'Lifetime Guarantee',
-      'limited' = 'One year warranty',
-      'as-is' = 'Sold as-is; no refunds or returns',
-   }
-
-   const product = getMockProduct({
-      variants: [
-         getMockProductVariant({
-            warranty: Warranty[guarantee],
-         }),
-      ],
-   });
-
-   return product;
 };
 
 export const getReviewsResponse = (
